% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_startdate_to_questions_for_session}
\alias{add_startdate_to_questions_for_session}
\title{Get the questions for a specific session id}
\usage{
add_startdate_to_questions_for_session(questions, session, sessionevents)
}
\arguments{
\item{questions}{questions to add startDate to}

\item{session}{The session to use to scan for \code{question-activated} events}

\item{sessionevents}{events that contain the \code{question-activated} events}
}
\value{
A \code{data.frame} with the session and question id and the \code{startDate} field
}
\description{
Goes through the associated slideshow, finds the questions ids and retrieves
the questions
}
