% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asq-multi-choice.R
\name{calc_mc_score}
\alias{calc_mc_score}
\title{Calculate the score for an answer of a multi-choice-question}
\usage{
calc_mc_score(submission, solution)
}
\arguments{
\item{submission}{JSON stringifed submission from a multi-choice answer}

\item{solution}{a vector of multi-choice-q options}
}
\value{
a vector with the parsed options
}
\description{
If the submission has exactly the same checked options as the solution
then the score is 100. Otherwise, it's 0.
}
