% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{calc_user_state}
\alias{calc_user_state}
\title{Calculate the user state from user events.}
\usage{
calc_user_state(timeline, user_events, from, to)
}
\arguments{
\item{timeline}{\code{date.frame} Each row represents a second in the lecture.}

\item{user_events}{\code{date.frame} Events from a \strong{single} user during the lecture.}

\item{from}{\code{POSIXct} pick user events that happend since the value of this variable.}

\item{to}{\code{POSIXct} pick user events that happend up to the value of this variable.}
}
\value{
An updated timeline \code{data.table} with the user states.
}
\description{
Given a \code{date.frame} timeline where each subsequent row corresponds to
the next a second of a lecture and the events for a \strong{single} user, it
calculates the state of the user for each second of the lecture as 6
different indicators:
}
\details{
\code{hasExercise}, \code{connected}, \code{visible}, \code{focus},
\code{idle}, \code{input} and \code{submitted}.
}

