% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{calc_user_state_for_users}
\alias{calc_user_state_for_users}
\title{Calculate user state for a bunch of users.}
\usage{
calc_user_state_for_users(users, events, startDate, endDate, slide_transitions)
}
\arguments{
\item{users}{\code{date.frame} Users from a lecture.}

\item{events}{\code{date.frame} Events from \strong{single} users during the lecture.}

\item{startDate}{\code{POSIXct} Calculate state for timestamps since this date.}

\item{endDate}{\code{POSIXct} Calculate state for timestamps up to this date.}

\item{slide_transitions}{\code{data.frame} Contains slide transitions.}
}
\value{
A \code{data.table} with the concatenated timelinesof all the users.
}
\description{
Goes over the \code{events} and \code{slide_transitions} to produce a
timeline with one second intervals from \code{startDate} to \code{endDate}
that holds 6 state indicators: \code{hasExercise}, \code{connected},
\code{focus}, \code{idle}, \code{input} and \code{submitted}.
}

