% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{calc_y_min_max_for_ribbon_plot}
\alias{calc_y_min_max_for_ribbon_plot}
\title{calculate the y min y max for each group (by state)}
\usage{
calc_y_min_max_for_ribbon_plot(time_dt)
}
\description{
intended to be used in a ribbon plot
}
