% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{create_time_datatable}
\alias{create_time_datatable}
\title{Create an empty timeline}
\usage{
create_time_datatable(from, to, slide_transitions)
}
\arguments{
\item{from}{\code{POSIXct} First timestamp of the timeline.}

\item{to}{\code{POSIXct} Second timestamp of the timeline.}

\item{slide_transitions}{\code{data.frame} Contains slide transitions.}
}
\value{
A \code{data.table} with the concatenated timelinesof all the users.
}
\description{
Goes over the \code{events} and \code{slide_transitions} to produce a
timeline with one second intervals from \code{startDate} to \code{endDate}
that holds 6 state indicators: \code{hasExercise}, \code{connected},
\code{focus}, \code{idle}, \code{input} and \code{submitted}.
}

