% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_activity_bar_time_faceted}
\alias{draw_activity_bar_time_faceted}
\title{stacked bar of user activity over time}
\usage{
draw_activity_bar_time_faceted(dt, slide_transitions_with_ex,
  slide_transitions_without_ex, state_labels)
}
\description{
stacked bar of user activity over time
}
