% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_activity_stacked_area}
\alias{draw_activity_stacked_area}
\title{stacked area plot}
\usage{
draw_activity_stacked_area(dt, slide_transitions_with_ex,
  slide_transitions_without_ex, state_labels, color_scale = get_color_scale(),
  num_students = 0, show_exercise_slide_overlay = TRUE,
  draw_states = TRUE, date_format = "\%H:\%M", date_break_secs = 300)
}
\description{
stacked area plot
}
