% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_activity_stacked_ribbon_by_slide}
\alias{draw_activity_stacked_ribbon_by_slide}
\title{stacked ribbon plot}
\usage{
draw_activity_stacked_ribbon_by_slide(dt, state_labels)
}
\description{
stacked ribbon plot
}

