% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_attention_focus_variation}
\alias{draw_attention_focus_variation}
\title{TODO: document}
\usage{
draw_attention_focus_variation(dt, slide_transitions_with_ex,
  slide_transitions_without_ex)
}
\description{
TODO: document
}

