% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{draw_connected_focus_area}
\alias{draw_connected_focus_area}
\title{stacked area plot}
\usage{
draw_connected_focus_area(dt, slide_transitions_with_ex,
  slide_transitions_without_ex, state_labels, color_scale = get_color_scale())
}
\description{
stacked area plot
}

