% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_questions_from_slideshow_questions_field}
\alias{extract_questions_from_slideshow_questions_field}
\title{Get a vector with the oids of the questions from the
JSON stringified `questions` field of a slideshow
document}
\usage{
extract_questions_from_slideshow_questions_field(questions)
}
\arguments{
\item{questions}{A JSON stringified string of the array of questions of a slideshow}
}
\value{
A vector of question oids
}
\description{
Get a vector with the oids of the questions from the
JSON stringified `questions` field of a slideshow
document
}
