% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_data_for_session}
\alias{filter_data_for_session}
\title{Filter data for a specific session.}
\usage{
filter_data_for_session(session, whitelistentries, sessionevents,
  slide_transitions)
}
\arguments{
\item{session}{the session to filter for}

\item{whitelistentries}{a data.frame with all the whitelistentries}

\item{sessionevents}{a data.frame with the all the sessionevents}

\item{slide_transitions}{a data.frame with the all the slide transitions}
}
\value{
A \code{list} that contains the \code{sessionevents}, \code{slide_transitions},
\code{slide_transitions_with_ex} (slide transitions with exercises),
\code{slide_transitions_without_ex} (slide transitions without exercises) and the \code{users}
for the specified session
}
\description{
This function is used to filter whitelist entries (users), session events
and slide transsionts for a specific ASQ session.
}
\examples{
session <- data.frame(id = c("s1"))
session[,1] <- factor(session[,1], levels=c("s1", "s2"))

whitelistentries = data.frame(
  id = c("wl1", "wl2", "wl3", "wl4", "wl5"),
  session = c("s1", "s2", "s1", "s1", "s2"),
  screenName = c("Fernscar Pirate", "Glazebard Dolphin", "Deepmouse Wolverine", "Planetpegasus Eater", "Crazyfalcon Runner"),
  role = c("presenter", "presenter", "viewer", "viewer", "viewer")
)

sessionevents = data.frame(
  session = c("s1", "s1", "s2"),
  type = c("type1", "type2", "type1")
)

slidetransitions = data.frame(
  session = c("s1", "s1", "s1", "s2", "s2"),
  slide = c("slide1", "slide2", "slide3", "slide1", "slide2"),
  hasExercise = c(TRUE, FALSE, TRUE, TRUE, FALSE)
)

filter_data_for_session(session, whitelistentries, sessionevents, slidetransitions)
}

