% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_slide_which_contains_question}
\alias{find_slide_which_contains_question}
\title{Get the slide that contains a question with a specific id}
\usage{
find_slide_which_contains_question(slideshow, qid)
}
\arguments{
\item{slideshow}{slideshow that contains the question}

\item{qid}{the question id}
}
\value{
the name of the slide
}
\description{
Goes through the associated slideshow and retrieves the name (HTML id)
of the slide that contains the question with the specified id.
}
