% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{fix_inconsistencies}
\alias{fix_inconsistencies}
\title{Fixes inconsistencies between states.}
\usage{
fix_inconsistencies(state_timeline_dt)
}
\arguments{
\item{state_timeline_dt}{\code{data.table} the state timeline data.table.}
}
\value{
An updated timeline \code{table} with the consistent states.
}
\description{
Goes through the 7 states: \code{hasExercise}, \code{connected}, \code{visible},
\code{focus}, \code{idle}, \code{input} and \code{submitted} and repairs
any inconsistencies. For example, if the user is not connected the rest of
the states should be \code{NA}.
}

