% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{format_timeline_for_state_time_plotting}
\alias{format_timeline_for_state_time_plotting}
\title{Prepare state timeline data for stacked area plotting}
\usage{
format_timeline_for_state_time_plotting(dt)
}
\arguments{
\item{dt}{timeline data.table with state information}
}
\value{
a data.table appropriate for the stacked plot
}
\description{
Groups by state and time, adds entries with count of 0 for the
states that don't appear in the data, calculates the height of
each ribbon and reorders the states
}
