% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asq-multi-choice.R
\name{get_mc_answers_with_score_for_questions}
\alias{get_mc_answers_with_score_for_questions}
\title{Get answers for the provided asq-multi-choice-questions and calculate their score}
\usage{
get_mc_answers_with_score_for_questions(answers, questions)
}
\arguments{
\item{answers}{a data.frame of answers to search in}

\item{questions}{a data.frame with questions to search answers for}
}
\value{
a data.frame with the matching answers and their scores
}
\description{
Get answers for the provided asq-multi-choice-questions and calculate their score
}
