% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asq-multi-choice.R
\name{get_mc_questions_for_session}
\alias{get_mc_questions_for_session}
\title{Get the multi-choice questions for a specific session id}
\usage{
get_mc_questions_for_session(sessions, slideshows, questions, sessionid)
}
\arguments{
\item{sessions}{sessions to search for the session with `sessionid`}

\item{slideshows}{slideshows to search for the matching slideshow}

\item{questions}{questions to search for multi-choice questions}

\item{sessionid}{the id of the session we want to find multi-choice questions for}
}
\value{
A data.frame with multi-choice questions
}
\description{
Goes through the associated slideshow, finds the questions ids, retrives
the questions and then filters for the `asq-multi-choice-q` type.
}

