% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asq-multi-choice.R
\name{get_n_back_timeline_with_correctness_for_mc}
\alias{get_n_back_timeline_with_correctness_for_mc}
\title{TODO: Document}
\usage{
get_n_back_timeline_with_correctness_for_mc(user.state, session, sessions,
  questions, answers, whitelistentries, sessionevents, slide.transitions)
}
\description{
TODO: Document
}
