% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_n_non_exercise_slides_before_question}
\alias{get_n_non_exercise_slides_before_question}
\title{Gets `n` non exercise slide transitions before the `question.slide`
plus the question.slide transition}
\usage{
get_n_non_exercise_slides_before_question(slide.transitions, session,
  question.slide, n = 3)
}
\arguments{
\item{slide.transitions}{a data.frame with the available slide.transitions that contain `question.slide`}

\item{session}{the session where the `question.slide` belongs}

\item{question.slide}{the name of the slide that contains the question of interest}

\item{n}{the number of consecutive non-exercise slides to search for}
}
\value{
a data.frame with the n+1 slide transitions
}
\description{
It searches for the longest slide transition for `question.slide` and then
it tries to find `n` consecutive slide transitions that don't contain
an exercise. It will return all the non-exercise slide transitions it
could find (up to `n`) plus the question.slide transition.
}
