% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_questions_for_session}
\alias{get_questions_for_session}
\title{Get the questions for a specific session id}
\usage{
get_questions_for_session(sessions, slideshows, questions, sessionid)
}
\arguments{
\item{sessions}{sessions to search for the session with `sessionid`}

\item{slideshows}{slideshows to search for the matching slideshow}

\item{questions}{questions to search}

\item{sessionid}{the id of the session we want to find questions for}
}
\value{
A data.frame with the found questions
}
\description{
Goes through the associated slideshow, finds the questions ids and retrieves
the questions
}

