% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{group_by_user_state}
\alias{group_by_user_state}
\title{group by user and state and add fields for pie charts}
\usage{
group_by_user_state(data)
}
\description{
group by user and state and add fields for pie charts
}
