% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{pad_timeline_with_0_count_states}
\alias{pad_timeline_with_0_count_states}
\title{TODO: document}
\usage{
pad_timeline_with_0_count_states(time_dt)
}
\description{
for each second that there is no observation for, between the min and
max time of timestamps in time_dt, inject observations with 0 count.
This stops plots from interpolating data.
}

