% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asq-multi-choice.R
\name{parse_mc_options}
\alias{parse_mc_options}
\title{Parse JSON stringified options from an asq-multi-choice-q question `data`
field}
\usage{
parse_mc_options(data)
}
\arguments{
\item{data}{JSON stringifed data-field of an asq-multi-choice-q question}
}
\value{
a vector with the parsed options
}
\description{
Parse JSON stringified options from an asq-multi-choice-q question `data`
field
}

