% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{prepare_data_for_state_plots}
\alias{prepare_data_for_state_plots}
\title{Prepare date for state plots.}
\usage{
prepare_data_for_state_plots(dt)
}
\arguments{
\item{dt}{\code{data.table} user state}
}
\value{
A \code(list) with
\itemize{
 \item{\code{dt}}{ The original \code{dt}}
 \item{\code{by_user_state_time_dt}}{ Per user state over time}
 \item{\code{by_state_time_dt}}{ State over time.}
 \item{\code{by_state_time_slide_dt}}{ State per slide over time.}
}
}
\description{
Utility function that reshapes user state data in the
shape needed by the plotting functions. Shapes:
\itemize{
 \item{\strong{Per user state over time}.}{ Calls \code{\link{group_by_user_state}}.}
 \item{\strong{State over time}.}{ Calls \code{\link{group_by_state_time}}.}
 \item{\strong{State per slide over time}.}{ Calls \code{\link{group_by_state_time_slide}}.}
}
}

