% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_questions_for_session}
\alias{get_questions_for_session}
\title{Get the questions for a specific session id}
\usage{
get_questions_for_session(session, slideshows, questions)
}
\arguments{
\item{session}{The session for which to get questions}

\item{slideshows}{slideshows to search for the matching slideshow}

\item{questions}{questions to search}
}
\value{
A data.frame with the found questions
}
\description{
Goes through the associated slideshow, finds the questions ids and retrieves
the questions
}

